src/requirements.txt
pandas
s3fs
psycopg2-binary
sqlalchemy
